+++ modules/back-end/src/Infrastructure/HealthCheckBuilderExtensions.cs
@@ -1,4 +1,5 @@
-﻿using Infrastructure.MQ;
+﻿using Infrastructure.Caches;
+using Infrastructure.MQ;
 using Infrastructure.MQ.Kafka;
 using Infrastructure.Persistence;
 using Microsoft.Extensions.Configuration;
@@ -38,13 +39,16 @@ public static class HealthCheckBuilderExtensions
             builder.AddKafka(configuration, tags, Timeout);
         }
 
-        // add redis health check
-        var redisConnectionString = configuration.GetValue<string>("Redis:ConnectionString")!;
-        builder.AddRedis(
-            redisConnectionString,
-            tags: tags,
-            timeout: Timeout
-        );
+        var cacheProvider = configuration.GetCacheProvider();
+        if (cacheProvider == CacheProvider.Redis)
+        {
+            var redisConnectionString = configuration.GetValue<string>("Redis:ConnectionString")!;
+            builder.AddRedis(
+                redisConnectionString,
+                tags: tags,
+                timeout: Timeout
+            );
+        }
 
         return builder;
     }
